+++ package.json
@@ -1,12 +1,12 @@
 {
   "name": "restapi-typescript-decorators",
-  "version": "3.2.4",
+  "version": "4.0.1",
   "description": "Decorators that helps reduce the boilderplate code for rest api calls on the frontend and node js",
   "main": "build/index.js",
   "types": "build/index.d.ts",
   "scripts": {
-    "build": "npx jest --clearCache && rm -rf build/* && ./node_modules/.bin/tsc",
-    "format": "./node_modules/.bin/prettier --config ./.prettierrc --write src/**/**/**/**/*.{ts,tsx,js,jsx,scss} *.{json,MD}",
+    "build": "rm -rf build/* && tsc",
+    "format": "npx prettier --config ./.prettierrc --write src/**/**/**/**/*.{ts,tsx,js,jsx,scss} *.{json,MD}",
     "clean": "npx jest --clearCache && rm -rf build/*",
     "test": "npx jest --ci"
   },
@@ -36,13 +36,13 @@
     "qs": "^6.9.4"
   },
   "devDependencies": {
-    "@types/node": "^14.0.22",
     "@types/jest": "^26.0.4",
+    "@types/node": "^14.0.23",
     "jest": "^26.1.0",
     "prettier": "1.19.1",
     "reflect-metadata": "^0.1.13",
     "ts-jest": "^26.1.1",
-    "typescript": "^3.8.3"
+    "typescript": "^3.9.6"
   },
   "peerDependencies": {}
 }
