--- src/__tests__/HttpBinTypes.ts
@@ -1,17 +1,0 @@
-export interface HttpBinAuthResponse {
-  authenticated: boolean;
-  user?: string;
-  token?: string;
-}
-
-export interface HttpBinGetResponse {
-  args?: object;
-  headers?: object;
-  origin?: string;
-  url?: string;
-  data?: object;
-  json?: string | object;
-  form?: object;
-}
-
-export interface HttpBinPostResponse extends HttpBinGetResponse {}
